refactor(fft)!: Move FFT autograd from coeus-autograd into coeus-fft
Apollo owns FFT; Coeus implements the autograd for Apollo. Repurpose the
coeus-fft crate from the redundant non-autograd tensor wrapper (#120) into the
dedicated home for that autograd, and remove FFT from coeus-autograd so every
FFT symbol has a single definition.

- coeus-fft now depends on coeus-autograd (Var/BackwardNode/GradBuffer) and
  apollo-fft (core fft_1d_slice_typed), owning the tensor-level fft_1d/ifft_1d,
  the Var-level fft_1d_var/ifft_1d_var/fft_energy, and the Fft1DNode/Ifft1DNode
  backward nodes. The one internal helper (should_track_var) is inlined to the
  public is_grad_enabled() + var.grad.
- Removed the fft module, its re-exports, and its tests from coeus-autograd;
  the FFT autograd tests move to coeus-fft/tests/fft.rs.
- coeus-python FFT bindings (ComplexTensor, fft/ifft/fft_energy) now source
  from coeus-fft. No change to the Python-visible surface.

Supersedes the G-FFT-CONSOLIDATE inversion plan.

BREAKING CHANGE: coeus_autograd::{fft_1d, ifft_1d, fft_1d_var, ifft_1d_var,
fft_energy, Fft1DNode, Ifft1DNode, FftScalar} move to coeus_fft. Internal-only:
the sole in-tree consumer (coeus-python) is updated; the Python API is unchanged.

Verification: coeus-fft + coeus-autograd + coeus-python chain builds clean;
coeus-fft tests 3/3 pass (analytic DFT + roundtrip, complex-seed input gradient,
Parseval energy-gradient oracle); fmt + clippy -D warnings clean.
