+++ coeus-autograd/src/lib.rs
@@ -80,9 +80,6 @@ pub use ops::{
     embedding,
     embedding_with_padding_idx,
     exp,
-    fft_1d,
-    fft_1d_var,
-    fft_energy,
     flip,
     floor,
     fold1d,
@@ -95,8 +92,6 @@ pub use ops::{
     hardswish,
     hardtanh,
     huber_loss,
-    ifft_1d,
-    ifft_1d_var,
     index_select,
     kl_divergence,
     l1_loss,
@@ -183,9 +178,6 @@ pub use ops::{
     AttentionMask,
     BatchNormArgs,
     CausalMask,
-    Fft1DNode,
-    FftScalar,
-    Ifft1DNode,
     NullMask,
     VarScalarExt,
 };
