+++ coeus-fft/tests/fft.rs
@@ -1,4 +1,5 @@
-use coeus_autograd::{fft_1d, fft_1d_var, fft_energy, ifft_1d, Var};
+use coeus_autograd::Var;
+use coeus_fft::{fft_1d, fft_1d_var, fft_energy, ifft_1d};
 use coeus_core::{Complex, MoiraiBackend};
 use coeus_tensor::Tensor;
